+++ crates/core/src/config/mod.rs
@@ -58,18 +58,34 @@ pub struct Config {
 
     #[serde(default)]
     pub catalog_source: Option<catalog::Source>,
+
+    #[serde(default)]
+    pub auto_update: bool,
 }
 
 impl Config {
+    /// Returns a `PathBuf` to the flavor directory.
+    pub fn get_flavor_directory_for_flavor(&self, flavor: &Flavor, path: &PathBuf) -> PathBuf {
+        path.join(&flavor.folder_name())
+    }
+
+    /// Returns a `Option<PathBuf>` to the root directory of the Flavor.
+    pub fn get_root_directory_for_flavor(&self, flavor: &Flavor) -> Option<PathBuf> {
+        if let Some(flavor_dir) = self.wow.directories.get(flavor) {
+            Some(flavor_dir.parent().unwrap().to_path_buf())
+        } else {
+            None
+        }
+    }
+
     /// Returns a `Option<PathBuf>` to the directory containing the addons.
     /// This will return `None` if no `wow_directory` is set in the config.
     pub fn get_addon_directory_for_flavor(&self, flavor: &Flavor) -> Option<PathBuf> {
-        match &self.wow.directory {
-            Some(wow_dir) => {
-                // The path to the flavor directory
-                let flavor_dir = wow_dir.join(&flavor.folder_name());
+        let dir = self.wow.directories.get(flavor);
+        match dir {
+            Some(dir) => {
                 // The path to the addons directory
-                let mut addon_dir = flavor_dir.join("Interface/AddOns");
+                let mut addon_dir = dir.join("Interface/AddOns");
 
                 // If path doesn't exist, it could have been modified by the user.
                 // Check for a case-insensitive version and use that instead.
@@ -81,7 +97,7 @@ impl Config {
 
                     // For some reason the case insensitive pattern doesn't work
                     // unless we add an actual pattern symbol, hence the `?`.
-                    let pattern = format!("{}/?nterface/?ddons", flavor_dir.display());
+                    let pattern = format!("{}/?nterface/?ddons", dir.display());
 
                     for entry in glob::glob_with(&pattern, options).unwrap() {
                         if let Ok(path) = entry {
@@ -93,7 +109,7 @@ impl Config {
                 // If flavor dir exists but not addon dir we try to create it.
                 // This state can happen if you do a fresh install of WoW and
                 // launch Ajour before you launch WoW.
-                if flavor_dir.exists() && !addon_dir.exists() {
+                if dir.exists() && !addon_dir.exists() {
                     let _ = create_dir_all(&addon_dir);
                 }
 
@@ -105,25 +121,19 @@ impl Config {
 
     /// Returns a `Option<PathBuf>` to the directory which will hold the
     /// temporary zip archives.
-    /// This will return `None` if no `wow_directory` is set in the config.
+    /// This will return `None` if flavor does not have a directory.
     pub fn get_download_directory_for_flavor(&self, flavor: Flavor) -> Option<PathBuf> {
-        match self.get_addon_directory_for_flavor(&flavor) {
-            Some(dir) => {
-                // The path to the directory which hold the temporary zip archives
-                let dir = dir.parent().expect("Expected Addons folder has a parent.");
-                Some(dir.to_path_buf())
-            }
-            None => None,
-        }
+        self.wow.directories.get(&flavor).cloned()
     }
 
     /// Returns a `Option<PathBuf>` to the WTF directory.
     /// This will return `None` if no `wow_directory` is set in the config.
     pub fn get_wtf_directory_for_flavor(&self, flavor: &Flavor) -> Option<PathBuf> {
-        match &self.wow.directory {
+        let dir = self.wow.directories.get(flavor);
+        match dir {
             Some(dir) => {
                 // The path to the WTF directory
-                let mut addon_dir = dir.join(&flavor.folder_name()).join("WTF");
+                let mut addon_dir = dir.join("WTF");
 
                 // If path doesn't exist, it could have been modified by the user.
                 // Check for a case-insensitive version and use that instead.
@@ -135,7 +145,7 @@ impl Config {
 
                     // For some reason the case insensitive pattern doesn't work
                     // unless we add an actual pattern symbol, hence the `?`.
-                    let pattern = format!("{}/?tf", dir.join(&flavor.folder_name()).display());
+                    let pattern = format!("{}/?tf", dir.display());
 
                     for entry in glob::glob_with(&pattern, options).unwrap() {
                         if let Ok(path) = entry {